Split self type validation
